move ImagePainter, and fix mobile drawImage quality
