+++ flutter/lib/utils/image.dart
@@ -1,6 +1,8 @@
 import 'dart:typed_data';
 import 'dart:ui' as ui;
 
+import 'package:flutter/widgets.dart';
+
 Future<ui.Image> decodeImageFromPixels(
   Uint8List pixels,
   int width,
@@ -46,4 +48,41 @@ Future<ui.Image> decodeImageFromPixels(
   buffer.dispose();
   descriptor.dispose();
   return frameInfo.image;
+}
+
+class ImagePainter extends CustomPainter {
+  ImagePainter({
+    required this.image,
+    required this.x,
+    required this.y,
+    required this.scale,
+  });
+
+  ui.Image? image;
+  double x;
+  double y;
+  double scale;
+
+  @override
+  void paint(Canvas canvas, Size size) {
+    if (image == null) return;
+    if (x.isNaN || y.isNaN) return;
+    canvas.scale(scale, scale);
+    // https://github.com/flutter/flutter/issues/76187#issuecomment-784628161
+    // https://api.flutter-io.cn/flutter/dart-ui/FilterQuality.html
+    var paint = Paint();
+    if ((scale - 1.0).abs() > 0.001) {
+      paint.filterQuality = FilterQuality.medium;
+      if (scale > 10.00000) {
+        paint.filterQuality = FilterQuality.high;
+      }
+    }
+    canvas.drawImage(
+        image!, Offset(x.toInt().toDouble(), y.toInt().toDouble()), paint);
+  }
+
+  @override
+  bool shouldRepaint(CustomPainter oldDelegate) {
+    return oldDelegate != this;
+  }
 }
